[types] Extract randomness DKG types into separate module
Refactors the DKG types module by extracting randomness-related types
(DKGTrait, DKGTranscript, DKGSessionMetadata, etc.) into a dedicated
randomness_dkg.rs file. This prepares for adding Chunky DKG types in
a separate module.
